+++ examples/category.mjs
@@ -1,7 +1,9 @@
+const constants = await (await fetch("/flux-ilias-rest-api/constants")).json();
+
 await (await fetch("/flux-ilias-rest-api/categories")).json();
 
 const time = Date.now();
-const category = await (await fetch(`/flux-ilias-rest-api/category/create/to-ref-id/${root.ref_id}`, {
+const category = await (await fetch(`/flux-ilias-rest-api/category/create/to-ref-id/${constants.root_object_ref_id}`, {
     method: "POST",
     headers: {
         "Content-Type": "application/json"
@@ -14,10 +16,9 @@ const category = await (await fetch(`/flux-ilias-rest-api/category/create/to-ref
 await (await fetch(`/flux-ilias-rest-api/category/by-id/${category.id}`)).json();
 
 await (await fetch(`/flux-ilias-rest-api/category/by-id/${category.id}/update`, {
-    method: "POST",
+    method: "PATCH",
     headers: {
-        "Content-Type": "application/json",
-        "X-Http-Method-Override": "PATCH"
+        "Content-Type": "application/json"
     },
     body: JSON.stringify({
         description: "Some description of the category"