+++ examples/course.mjs
@@ -1,7 +1,9 @@
+const constants = await (await fetch("/flux-ilias-rest-api/constants")).json();
+
 await (await fetch("/flux-ilias-rest-api/courses")).json();
 
 const time = Date.now();
-const course = await (await fetch(`/flux-ilias-rest-api/course/create/to-ref-id/${root.ref_id}`, {
+const course = await (await fetch(`/flux-ilias-rest-api/course/create/to-ref-id/${constants.root_object_ref_id}`, {
     method: "POST",
     headers: {
         "Content-Type": "application/json"
@@ -16,10 +18,9 @@ const course = await (await fetch(`/flux-ilias-rest-api/course/create/to-ref-id/
 await (await fetch(`/flux-ilias-rest-api/course/by-id/${course.id}`)).json();
 
 await (await fetch(`/flux-ilias-rest-api/course/by-id/${course.id}/update`, {
-    method: "POST",
+    method: "PATCH",
     headers: {
-        "Content-Type": "application/json",
-        "X-Http-Method-Override": "PATCH"
+        "Content-Type": "application/json"
     },
     body: JSON.stringify({
         description: "Some description of the course",