+++ examples/group.mjs
@@ -1,7 +1,9 @@
+const constants = await (await fetch("/flux-ilias-rest-api/constants")).json();
+
 await (await fetch("/flux-ilias-rest-api/groups")).json();
 
 const time = Date.now();
-const group = await (await fetch(`/flux-ilias-rest-api/group/create/to-ref-id/${root.ref_id}`, {
+const group = await (await fetch(`/flux-ilias-rest-api/group/create/to-ref-id/${constants.root_object_ref_id}`, {
     method: "POST",
     headers: {
         "Content-Type": "application/json"
@@ -14,10 +16,9 @@ const group = await (await fetch(`/flux-ilias-rest-api/group/create/to-ref-id/${
 await (await fetch(`/flux-ilias-rest-api/group/by-id/${group.id}`)).json();
 
 await (await fetch(`/flux-ilias-rest-api/group/by-id/${group.id}/update`, {
-    method: "POST",
+    method: "PATCH",
     headers: {
-        "Content-Type": "application/json",
-        "X-Http-Method-Override": "PATCH"
+        "Content-Type": "application/json"
     },
     body: JSON.stringify({
         description: "Some description of the group"