+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-04-26-1
+
+Changes:
+
+- Remove method override fallback
+
 ## v2023-04-24-1
 
 Changes: